% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{assign_gene_symbols}
\alias{assign_gene_symbols}
\title{Assign gene symbols to peak lists}
\usage{
assign_gene_symbols(peak_list, contrast_full)
}
\arguments{
\item{peak_list}{Character string of Entrez IDs}

\item{contrast_full}{Character string contrast name}
}
\value{
Character string of gene symbols
}
\description{
Internal helper function for enrichment analysis
}
\keyword{internal}
