% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacreportR-package.R
\docType{package}
\name{atacreportR-package}
\alias{atacreportR}
\alias{atacreportR-package}
\title{atacreportR: ATAC-seq Differential Peak Analysis and Reporting Functions}
\description{
Core functions for ATAC-seq differential accessibility analysis, including peak processing, count normalization, statistical testing with edgeR, annotation, and report generation. Designed for use with the UFHCC BCB-SR ATAC-seq analysis pipeline.
}
\author{
\strong{Maintainer}: Heather Kates \email{hkates@ufl.edu} (\href{https://orcid.org/0000-0001-8490-2201}{ORCID})

}
\keyword{internal}
