% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{clean_group_name}
\alias{clean_group_name}
\title{Clean group names by removing leading X prefix}
\usage{
clean_group_name(name)
}
\arguments{
\item{name}{Character string to clean}
}
\value{
Cleaned character string
}
\description{
Clean group names by removing leading X prefix
}
