% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{copy_file_to_web}
\alias{copy_file_to_web}
\title{Copy file to web directory}
\usage{
copy_file_to_web(source_path, dest_dir, filename)
}
\arguments{
\item{source_path}{Source file path}

\item{dest_dir}{Destination directory}

\item{filename}{Destination filename}
}
\value{
Logical indicating success
}
\description{
Copy file to web directory
}
