% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_processing.R
\name{create_saf_file}
\alias{create_saf_file}
\title{Create SAF format annotation file for featureCounts}
\usage{
create_saf_file(consensus_peaks)
}
\arguments{
\item{consensus_peaks}{GRanges object containing genomic regions}
}
\value{
Character string path to temporary SAF file
}
\description{
Converts a GRanges object to SAF (Simplified Annotation Format) for use
with Rsubread::featureCounts.
}
\keyword{internal}
