% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.add_bigwig_paths_to_dds}
\alias{.add_bigwig_paths_to_dds}
\title{Add BigWig file paths to DDS colData}
\usage{
.add_bigwig_paths_to_dds(dds, bigwig_files)
}
\description{
Add BigWig file paths to DDS colData
}
\keyword{internal}
