% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.add_peak_paths_to_dds}
\alias{.add_peak_paths_to_dds}
\title{Add peak file paths to DDS colData}
\usage{
.add_peak_paths_to_dds(dds, peak_files)
}
\description{
Add peak file paths to DDS colData
}
\keyword{internal}
