% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.find_contrast_variable}
\alias{.find_contrast_variable}
\title{Find which metadata column contains the contrast groups}
\usage{
.find_contrast_variable(sample_metadata, group1, group2)
}
\description{
Find which metadata column contains the contrast groups
}
\keyword{internal}
