% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.merge_peak_anno_to_dds}
\alias{.merge_peak_anno_to_dds}
\title{Merge peak annotations into DDS rowData}
\usage{
.merge_peak_anno_to_dds(dds, peaks_anno)
}
\description{
Merge peak annotations into DDS rowData
}
\keyword{internal}
