% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.merge_sample_info_to_dds}
\alias{.merge_sample_info_to_dds}
\title{Merge sample info into DDS colData}
\usage{
.merge_sample_info_to_dds(dds, sample_info)
}
\description{
Merge sample info into DDS colData
}
\keyword{internal}
