% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.resolve_output_dir}
\alias{.resolve_output_dir}
\title{Resolve output directory and create if needed}
\usage{
.resolve_output_dir(report_params)
}
\description{
Resolve output directory and create if needed
}
\keyword{internal}
