% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.run_single_contrast}
\alias{.run_single_contrast}
\title{Run differential analysis for a single contrast}
\usage{
.run_single_contrast(dge, contrast_pair, min_count, min_prop)
}
\description{
Run differential analysis for a single contrast
}
\keyword{internal}
