% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.write_consensus_peaks}
\alias{.write_consensus_peaks}
\title{Write consensus peaks to BED-like text file}
\usage{
.write_consensus_peaks(peaks_gr, out_file)
}
\description{
Write consensus peaks to BED-like text file
}
\keyword{internal}
