% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.write_peak_annotation}
\alias{.write_peak_annotation}
\title{Write peak annotation to file}
\usage{
.write_peak_annotation(peak_anno, out_path)
}
\description{
Write peak annotation to file
}
\keyword{internal}
