% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{download_button_png}
\alias{download_button_png}
\title{Create downloadable PNG button for plots}
\usage{
download_button_png(
  plot_object,
  output_name = "plot",
  width = 12,
  height = 6,
  dpi = 300
)
}
\arguments{
\item{plot_object}{ggplot2 plot object}

\item{output_name}{Character string for filename}

\item{width}{Numeric plot width in inches}

\item{height}{Numeric plot height in inches}

\item{dpi}{Numeric resolution}
}
\value{
HTML download button
}
\description{
Create downloadable PNG button for plots
}
