% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_processing.R
\name{match_sample_info_to_counts}
\alias{match_sample_info_to_counts}
\title{Match sample info to count matrix columns}
\usage{
match_sample_info_to_counts(sample_info, count_matrix)
}
\arguments{
\item{sample_info}{Data frame with sample metadata}

\item{count_matrix}{Matrix of counts}
}
\value{
Data frame with sample info matched to count matrix
}
\description{
Match sample info to count matrix columns
}
\keyword{internal}
