% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Fallback value}
}
\value{
Either x or y
}
\description{
Returns y if x is NULL or empty string, otherwise returns x
}
\keyword{internal}
