% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_funcs.R
\name{preprocess_result}
\alias{preprocess_result}
\title{Preprocess differential accessibility results for visualization}
\usage{
preprocess_result(df, contrast)
}
\arguments{
\item{df}{Data frame with DA results}

\item{contrast}{Character string contrast name}
}
\value{
Data frame with added visualization columns
}
\description{
Adds grouping and significance columns for plotting.
}
