% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{process_nfcore_format}
\alias{process_nfcore_format}
\title{Process nf-core format sample sheet}
\usage{
process_nfcore_format(sample_info)
}
\description{
Process nf-core format sample sheet
}
\keyword{internal}
