% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{process_simple_format}
\alias{process_simple_format}
\title{Process simple format sample sheet}
\usage{
process_simple_format(sample_info)
}
\description{
Process simple format sample sheet
}
\keyword{internal}
